+++ src/Utilities/Parser/AlbumesParser.ts
@@ -1,4 +1,4 @@
-import {Album} from "../bd/entity/Album";
+import {Album} from "../../bd/entity/Album";
 
 
 export class AlbumParser{
