+++ src/bd/controllersBd/AlbumesRepository.ts
@@ -1,6 +1,6 @@
 import {getConnection,getRepository,getConnectionManager,createConnection} from "typeorm";
 import {Album} from "../entity/Album";
-import {AlbumParser} from "../../Parser/AlbumesParser";
+import {AlbumParser} from "../../Utilities/Parser/AlbumesParser";
 import {Cancion} from "../entity/Cancion";
 import {Artista} from "../entity/Artista";
 import {v4 as uuidv4} from "uuid";
